Add tests for EIS / GIS

+++ python/pspython/methods/techniques.py
@@ -498,12 +498,12 @@ class ElectrochemicalImpedanceSpectroscopyParameters(
         Minimum frequency in Hz (default: 1e3)
     """
 
-    equilibration_time: float = 0.0  # Time (s)
-    dc_potential: float = 0.0  # in V
-    ac_potential: float = 0.01  # in V RMS
-    n_frequencies: float = 11  # Number of frequencies
-    max_frequency: float = 1e5  # in Hz
-    min_frequency: float = 1e3  # in Hz
+    equilibration_time: float = 0.0
+    dc_potential: float = 0.0
+    ac_potential: float = 0.01
+    n_frequencies: int = 11
+    max_frequency: float = 1e5
+    min_frequency: float = 1e3
 
     _PSMethod = Techniques.ImpedimetricMethod
 
@@ -551,13 +551,13 @@ class GalvanostaticImpedanceSpectroscopyParameters(
         Minimum frequency in Hz (default: 1e3)
     """
 
-    applied_current_range: float = get_current_range(6)  # in applied current range
-    equilibration_time: float = 0.0  # Time (s
-    ac_current: float = 0.01  # in applied current range RMS
-    dc_current: float = 0.0  # in applied current range
-    n_frequencies: int = 11  # Number of frequencies
-    max_frequency: float = 1e5  # in Hz
-    min_frequency: float = 1e3  # in Hz
+    applied_current_range: float = get_current_range(6)
+    equilibration_time: float = 0.0
+    ac_current: float = 0.01
+    dc_current: float = 0.0
+    n_frequencies: int = 11
+    max_frequency: float = 1e5
+    min_frequency: float = 1e3
 
     _PSMethod = Techniques.ImpedimetricGstatMethod
